+++ janitor/functions/utils.py
@@ -642,9 +642,7 @@ def get_columns(group: Union[DataFrameGroupBy, SeriesGroupBy], label):
 
 def _select(
     df: pd.DataFrame,
-    args: tuple,
     invert: bool = False,
-    axis: str = "index",
     rows=None,
     columns=None,
 ) -> pd.DataFrame:
@@ -653,23 +651,25 @@ def _select(
 
     Returns a DataFrame.
     """
-    assert axis in {"both", "index", "columns"}
-    if axis == "both":
-        if rows is None:
-            rows = slice(None)
+    if rows is None:
+        row_indexer = slice(None)
+    else:
+        outcome = _select_index([rows], df, axis="index")
+        if invert:
+            row_indexer = np.ones(df.index.size, dtype=np.bool_)
+            row_indexer[outcome] = False
         else:
-            rows = _select_index([rows], df, axis="index")
-        if columns is None:
-            columns = slice(None)
+            row_indexer = outcome
+    if columns is None:
+        column_indexer = slice(None)
+    else:
+        outcome = _select_index([columns], df, axis="columns")
+        if invert:
+            column_indexer = np.ones(df.columns.size, dtype=np.bool_)
+            column_indexer[outcome] = False
         else:
-            columns = _select_index([columns], df, axis="columns")
-        return df.iloc[rows, columns]
-    indices = _select_index(list(args), df, axis)
-    if invert:
-        rev = np.ones(getattr(df, axis).size, dtype=np.bool_)
-        rev[indices] = False
-        return df.iloc(axis=axis)[rev]
-    return df.iloc(axis=axis)[indices]
+            column_indexer = outcome
+    return df.iloc[row_indexer, column_indexer]
 
 
 class _JoinOperator(Enum):